Crate pallet_subspace

Subspace consensus pallet.

This pallet is in many ways complementary to sc-consensus-subspace.

Pallet maintains crucial state required for Subspace Proof-of-Archival-Storage consensus:

global randomness

based on c-correlation from https://arxiv.org/abs/1910.02218

is used to later derive together with time slot a global challenge

solution range, which is a range of valid solution for Proof-of-Archival-Storage puzzle

conceptually similar to work difficulty in Proof-of-Work consensus

is updated every Era

salt, which is used for creating plot commitments and is updated every Eon

inherents for:

storing root blocks and maintaining mapping from segment index to corresponding records root such that validity of piece from solution can be checked later

handling of farmer equivocation (together with pallet-offences-subspace) and maintaining list of blocked farmers (effectively burned plots)

Pallet also provides handy API for finding block author, block reward address, randomness and some others.

License: Apache-2.0

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

equivocation

An opt-in utility module for reporting equivocations.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

NormalEonChange

A type signifying to Subspace that it should perform eon changes with an internal trigger.

NormalEraChange

A type signifying to Subspace that it should perform era changes with an internal trigger.

NormalGlobalRandomnessInterval

A type signifying to Subspace that it should perform a global randomness update with an internal trigger.

Pallet

The Subspace Pallet

SolutionRangeOverride

Override for next solution range adjustment

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

Events type.

Traits

Config

Configuration trait of this pallet.

EonChangeTrigger

Trigger an eon change, if any should take place.

EraChangeTrigger

Trigger an era change, if any should take place.

GlobalRandomnessIntervalTrigger

Trigger global randomness every interval.

WeightInfo

Type Definitions

CurrentSlot

Current slot number.

EonIndex

Current eon index.

EraStartSlot

Slot at which current era started.

GenesisSlot

The slot at which the block was created. This is 0 until the first block of the chain.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextSolutionRangeOverride

Override solution range during next update

Salts

Salts used for challenges.

ShouldAdjustSolutionRange

Storage to check if the solution range is to be adjusted for next era