Trait pallet_subspace::equivocation::HandleEquivocation

pub trait HandleEquivocation<T: Config> {
    type ReportLongevity: Get<u64>;
    fn report_offence(
        offence: SubspaceEquivocationOffence<FarmerPublicKey>
    ) -> Result<(), OffenceError>;

    fn is_known_offence(offenders: &[FarmerPublicKey], time_slot: &Slot) -> bool;

    fn submit_equivocation_report(
        equivocation_proof: EquivocationProof<T::Header>
    ) -> DispatchResult;
}

A trait with utility methods for handling equivocation reports in Subspace. The trait provides methods for reporting an offence triggered by a valid equivocation report, checking the current block author (to declare as the reporter), and also for creating and submitting equivocation report extrinsics (useful only in offchain context).

Associated Types

type ReportLongevity: Get<u64>

The longevity, in blocks, that the equivocation report is valid for. When using the staking pallet this should be equal to the bonding duration (in blocks, not eras).

Required methods

fn report_offence(
offence: SubspaceEquivocationOffence<FarmerPublicKey>
) -> Result<(), OffenceError>

Report an offence proved by the given reporters.

fn is_known_offence(offenders: &[FarmerPublicKey], time_slot: &Slot) -> bool

Returns true if all of the offenders at the given time slot have already been reported.

fn submit_equivocation_report(
equivocation_proof: EquivocationProof<T::Header>
) -> DispatchResult

Create and dispatch an equivocation report extrinsic.