pallet_subspace
Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Subspace Pallet

Enums

AllowAuthoringBy

Call

Contains a variant per dispatchable extrinsic that this pallet has.

EnableRewardsAt

When to enable block/vote rewards

Error

The Error enum of this pallet.

Event

Events type.

Traits

Config

Configuration trait of this pallet.

Type Aliases

BlockRandomness

The current block randomness, updated at block initialization. When the proof of time feature is enabled it derived from PoT otherwise PoR.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.