pallet_subspace::pallet::dispatchables

Function store_segment_headers

pub fn store_segment_headers<T: Config>(segment_headers: Vec<SegmentHeader>)

Submit new segment header to the blockchain. This is an inherent extrinsic and part of the Subspace consensus logic.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::store_segment_headers.