Struct pallet_subspace::Pallet

pub struct Pallet<T>(_);

The Subspace Pallet

Implementations

impl<T: Config> Pallet<T>

Methods for the ValidateUnsigned implementation: It restricts calls to report_equivocation to local calls (i.e. extrinsics generated on this node) or that already in a block. This guarantees that only block authors can include equivocation reports.

pub fn validate_equivocation_report(
source: TransactionSource,
equivocation_proof: &EquivocationProof<T::Header>
) -> TransactionValidity

pub fn pre_dispatch_equivocation_report(
equivocation_proof: &EquivocationProof<T::Header>
) -> Result<(), TransactionValidityError>

impl<T: Config> Pallet<T>

pub fn report_equivocation(
origin: OriginFor<T>,
equivocation_proof: Box<EquivocationProof<T::Header>>
) -> DispatchResultWithPostInfo

Report farmer equivocation/misbehavior. This method will verify the equivocation proof. If valid, the offence will be reported.

This extrinsic must be called unsigned and it is expected that only block authors will call it (validated in ValidateUnsigned), as such if the block author is defined it will be defined as the equivocation reporter.

pub fn store_root_blocks(
origin: OriginFor<T>,
root_blocks: Vec<RootBlock>
) -> DispatchResult

Submit new root block to the blockchain. This is an inherent extrinsic and part of the Subspace consensus logic.

pub fn enable_solution_range_adjustment(
origin: OriginFor<T>,
solution_range_override: Option<u64>,
voting_solution_range_override: Option<u64>
) -> DispatchResult

Enable solution range adjustment after every era. Note: No effect on the solution range for the current era

pub fn vote(
origin: OriginFor<T>,
signed_vote: Box<SignedVote<T::BlockNumber, T::Hash, T::AccountId>>
) -> DispatchResult

Farmer vote, currently only used for extra rewards to farmers.

pub fn enable_rewards(
origin: OriginFor<T>,
height: Option<T::BlockNumber>
) -> DispatchResult

Enable rewards for blocks and votes at specified block height.

pub fn enable_storage_access(origin: OriginFor<T>) -> DispatchResult

Enable storage access for all users.

pub fn enable_authoring_by_anyone(origin: OriginFor<T>) -> DispatchResult

Enable storage access for all users.

impl<T: Config> Pallet<T>

pub fn genesis_slot() -> Slot

The slot at which the first block was created. This is 0 until the first block of the chain.

impl<T: Config> Pallet<T>

pub fn current_slot() -> Slot

Current slot number.

impl<T: Config> Pallet<T>

pub fn global_randomnesses() -> GlobalRandomnesses

Global randomnesses derived from from PoR signature and used for deriving global challenges.

impl<T: Config> Pallet<T>

pub fn solution_ranges() -> SolutionRanges

Solution ranges used for challenges.

impl<T: Config> Pallet<T>

pub fn should_adjust_solution_range() -> bool

Storage to check if the solution range is to be adjusted for next era

impl<T: Config> Pallet<T>

pub fn records_root<KArg>(k: KArg) -> Option<RecordsRoot>where
KArg: EncodeLike<SegmentIndex>,

Mapping from segment index to corresponding merkle tree root of segment records.

impl<T: Config> Pallet<T>

pub fn is_storage_access_enabled() -> bool

Enable storage access for all users.

impl<T: Config> Pallet<T>

pub fn root_plot_public_key() -> Option<FarmerPublicKey>

Root plot public key.

Set just once to make sure no one else can author blocks until allowed for anyone.

impl<T: Config> Pallet<T>

pub fn slot_duration() -> T::Moment

Determine the Subspace slot duration based on the Timestamp module configuration.

pub fn total_pieces() -> NonZeroU64

Total number of pieces in the blockchain

pub fn submit_equivocation_report(
equivocation_proof: EquivocationProof<T::Header>
) -> Option<()>

Submits an extrinsic to report an equivocation. This method will create an unsigned extrinsic with a call to report_equivocation and will push the transaction to the pool. Only useful in an offchain context.

pub fn submit_test_equivocation_report(
equivocation_proof: EquivocationProof<T::Header>
) -> Option<()>

Just stores offender from equivocation report in block list, only used for tests.

pub fn is_in_block_list(farmer_public_key: &FarmerPublicKey) -> bool

Check if farmer_public_key is in block list (due to equivocation)

pub fn archived_history_size() -> u64

Size of the archived history of the blockchain in bytes

pub fn chain_constants() -> ChainConstants

impl<T> Pallet<T>where
T: Config + SendTransactionTypes<Call<T>>,

pub fn submit_vote(
signed_vote: SignedVote<T::BlockNumber, T::Hash, T::AccountId>
)

Submit farmer vote vote that is essentially a header with bigger solution range than acceptable for block authoring.