Trait pallet_subspace::Config

pub trait Config: Config {

    type Event: From<Event> + IsType<<Self as Config>::Event>;
    type EpochDuration: Get<u64>;
    type EraDuration: Get<u32>;
    type EonDuration: Get<u64>;
    type InitialSolutionRange: Get<u64>;
    type SlotProbability: Get<(u64, u64)>;
    type ExpectedBlockTime: Get<Self::Moment>;
    type ConfirmationDepthK: Get<u32>;
    type RecordSize: Get<u32>;
    type RecordedHistorySegmentSize: Get<u32>;
    type EpochChangeTrigger: EpochChangeTrigger;
    type EraChangeTrigger: EraChangeTrigger;
    type EonChangeTrigger: EonChangeTrigger;
    type HandleEquivocation: HandleEquivocation<Self>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event> + IsType<<Self as Config>::Event>

The overarching event type.

type EpochDuration: Get<u64>

The amount of time, in slots, that each epoch should last. NOTE: Currently it is not possible to change the epoch duration after the chain has started. Attempting to do so will brick block production.

type EraDuration: Get<u32>

The amount of time, in blocks, that each era should last. NOTE: Currently it is not possible to change the era duration after the chain has started. Attempting to do so will brick block production.

type EonDuration: Get<u64>

The amount of time, in slots, that each eon should last. NOTE: Currently it is not possible to change the eon duration after the chain has started. Attempting to do so will brick block production.

type InitialSolutionRange: Get<u64>

Initial solution range used for challenges during the very first era.

type SlotProbability: Get<(u64, u64)>

How often in slots slots (on average, not counting collisions) will have a block.

Expressed as a rational where the first member of the tuple is the numerator and the second is the denominator. The rational should represent a value between 0 and 1.

type ExpectedBlockTime: Get<Self::Moment>

The expected average block time at which Subspace should be creating blocks. Since Subspace is probabilistic it is not trivial to figure out what the expected average block time should be based on the slot duration and the security parameter c (where 1 - c represents the probability of a slot being empty).

type ConfirmationDepthK: Get<u32>

Depth K after which a block enters the recorded history (a global constant, as opposed to the client-dependent transaction confirmation depth k).

type RecordSize: Get<u32>

The size of data in one piece (in bytes).

type RecordedHistorySegmentSize: Get<u32>

Recorded history is encoded and plotted in segments of this size (in bytes).

type EpochChangeTrigger: EpochChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an epoch has ended and to perform the transition to the next epoch.

Typically, the ExternalTrigger type should be used. An internal trigger should only be used when no other module is responsible for changing epochs.

type EraChangeTrigger: EraChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an era has ended and to perform the transition to the next era.

type EonChangeTrigger: EonChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an eon has ended and to perform the transition to the next eon.

type HandleEquivocation: HandleEquivocation<Self>

The equivocation handling subsystem, defines methods to report an offence (after the equivocation has been validated) and for submitting a transaction to report an equivocation (from an offchain context). NOTE: when enabling equivocation handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.