Crate pallet_subspace

Subspace consensus pallet.

This pallet is in many ways complementary to sc-consensus-subspace.

Pallet maintains crucial state required for Subspace Proof-of-Archival-Storage consensus:

global randomness

based on c-correlation from https://arxiv.org/abs/1910.02218

is used to later derive together with time slot a global challenge

solution range, which is a range of valid solution for Proof-of-Archival-Storage puzzle

conceptually similar to work difficulty in Proof-of-Work consensus

is updated every Era

inherents for:

storing segment headers and maintaining mapping from segment index to corresponding segment commitment such that validity of piece from solution can be checked later

Pallet also provides handy API for finding block author, block reward address, randomness and some others.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

extensions

Extensions for unsigned general extrinsics

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_subspace

Structs

EnsureSubspaceOrigin

Ensure the subspace origin.

NormalEraChange

A type signifying to Subspace that it should perform era changes with an internal trigger.

Enums

RawOrigin

Custom origin for validated unsigned extrinsics.

Traits

EraChangeTrigger

Trigger an era change, if any should take place.