Module pallet_subspace::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

enable_authoring_by_anyone

Enable storage access for all users.

enable_rewards_at

Enable rewards for blocks and votes at specified block height.

enable_solution_range_adjustment

Enable solution range adjustment after every era. Note: No effect on the solution range for the current era

report_equivocation

Report farmer equivocation/misbehavior. This method will verify the equivocation proof. If valid, the offence will be reported.

store_segment_headers

Submit new segment header to the blockchain. This is an inherent extrinsic and part of the Subspace consensus logic.

vote

Farmer vote, currently only used for extra rewards to farmers.