pallet_subspace::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AllowAuthoringByAnyone

Allow block authoring by anyone or just root.

BlockRandomness

The current block randomness, updated at block initialization. When the proof of time feature is enabled it derived from PoT otherwise PoR.

BlockSlots

Bounded mapping from block number to slot

CurrentBlockAuthorInfo

Block author information

CurrentBlockVoters

Voters in the current block thus far

DidProcessSegmentHeaders

Whether the segment headers inherent has been processed in this block (temporary value).

EnableRewards

Enable rewards since specified block number.

EnableRewardsBelowSolutionRange

Enable rewards when solution range is below this threshold.

EraStartSlot

Slot at which current era started.

NextSolutionRangeOverride

Override solution range during next update

ParentBlockAuthorInfo

Parent block author information.

ParentBlockVoters

Voters in the parent block (set at the end of the block with current values).

ParentVoteVerificationData

Storage of previous vote verification data, updated on each block during finalization.

PotEntropy

Entropy that needs to be injected into proof of time chain at specific slot associated with block number it came from.

PotSlotIterations

Number of iterations for proof of time per slot with optional scheduled update

RootPlotPublicKey

Root plot public key.

SegmentCommitment

Mapping from segment index to corresponding segment commitment of contained records.

ShouldAdjustSolutionRange

Storage to check if the solution range is to be adjusted for next era

SolutionRanges

Solution ranges used for challenges.