Trait pallet_subspace::Config

pub trait Config: Config {

    type Event: From<Event> + IsType<<Self as Config>::Event>;
    type GlobalRandomnessUpdateInterval: Get<Self::BlockNumber>;
    type EraDuration: Get<Self::BlockNumber>;
    type EonDuration: Get<u64>;
    type EonNextSaltReveal: Get<u64>;
    type InitialSolutionRange: Get<u64>;
    type SlotProbability: Get<(u64, u64)>;
    type ExpectedBlockTime: Get<Self::Moment>;
    type ConfirmationDepthK: Get<Self::BlockNumber>;
    type RecordSize: Get<u32>;
    type MaxPlotSize: Get<u64>;
    type RecordedHistorySegmentSize: Get<u32>;
    type ShouldAdjustSolutionRange: Get<bool>;
    type GlobalRandomnessIntervalTrigger: GlobalRandomnessIntervalTrigger;
    type EraChangeTrigger: EraChangeTrigger;
    type EonChangeTrigger: EonChangeTrigger;
    type HandleEquivocation: HandleEquivocation<Self>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event> + IsType<<Self as Config>::Event>

The overarching event type.

type GlobalRandomnessUpdateInterval: Get<Self::BlockNumber>

The amount of time, in blocks, between updates of global randomness.

type EraDuration: Get<Self::BlockNumber>

The amount of time, in blocks, that each era should last. NOTE: Currently it is not possible to change the era duration after the chain has started. Attempting to do so will brick block production.

type EonDuration: Get<u64>

The amount of time, in slots, that each eon should last. NOTE: Currently it is not possible to change the eon duration after the chain has started. Attempting to do so will brick block production.

type EonNextSaltReveal: Get<u64>

The amount of time within eon, in slots, after which next eon salt should be revealed.

The purpose of this is to allow to start tag recommitment a bit upfront, but not too soon. For instance, if eon duration is 7 days, this parameter may be set to 6 days worth of timeslots.

type InitialSolutionRange: Get<u64>

Initial solution range used for challenges during the very first era.

type SlotProbability: Get<(u64, u64)>

How often in slots slots (on average, not counting collisions) will have a block.

Expressed as a rational where the first member of the tuple is the numerator and the second is the denominator. The rational should represent a value between 0 and 1.

type ExpectedBlockTime: Get<Self::Moment>

The expected average block time at which Subspace should be creating blocks. Since Subspace is probabilistic it is not trivial to figure out what the expected average block time should be based on the slot duration and the security parameter c (where 1 - c represents the probability of a slot being empty).

type ConfirmationDepthK: Get<Self::BlockNumber>

Depth K after which a block enters the recorded history (a global constant, as opposed to the client-dependent transaction confirmation depth k).

type RecordSize: Get<u32>

The size of data in one piece (in bytes).

type MaxPlotSize: Get<u64>

Maximum number of pieces in each plot

type RecordedHistorySegmentSize: Get<u32>

Recorded history is encoded and plotted in segments of this size (in bytes).

type ShouldAdjustSolutionRange: Get<bool>

type GlobalRandomnessIntervalTrigger: GlobalRandomnessIntervalTrigger

Subspace requires periodic global randomness update.

type EraChangeTrigger: EraChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an era has ended and to perform the transition to the next era.

Era is normally used to update solution range used for challenges.

type EonChangeTrigger: EonChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an eon has ended and to perform the transition to the next eon.

Era is normally used to update salt used for plot commitments.

type HandleEquivocation: HandleEquivocation<Self>

The equivocation handling subsystem, defines methods to report an offence (after the equivocation has been validated) and for submitting a transaction to report an equivocation (from an offchain context).

NOTE: when enabling equivocation handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.