Trait pallet_subspace::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type GlobalRandomnessUpdateInterval: Get<Self::BlockNumber>;
    type EraDuration: Get<Self::BlockNumber>;
    type InitialSolutionRange: Get<SolutionRange>;
    type SlotProbability: Get<(u64, u64)>;
    type ExpectedBlockTime: Get<Self::Moment>;
    type ConfirmationDepthK: Get<Self::BlockNumber>;
    type ExpectedVotesPerBlock: Get<u32>;
    type ShouldAdjustSolutionRange: Get<bool>;
    type GlobalRandomnessIntervalTrigger: GlobalRandomnessIntervalTrigger;
    type EraChangeTrigger: EraChangeTrigger;
    type HandleEquivocation: HandleEquivocation<Self>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type GlobalRandomnessUpdateInterval: Get<Self::BlockNumber>

The amount of time, in blocks, between updates of global randomness.

type EraDuration: Get<Self::BlockNumber>

The amount of time, in blocks, that each era should last. NOTE: Currently it is not possible to change the era duration after the chain has started. Attempting to do so will brick block production.

type InitialSolutionRange: Get<SolutionRange>

Initial solution range used for challenges during the very first era.

type SlotProbability: Get<(u64, u64)>

How often in slots slots (on average, not counting collisions) will have a block.

Expressed as a rational where the first member of the tuple is the numerator and the second is the denominator. The rational should represent a value between 0 and 1.

type ExpectedBlockTime: Get<Self::Moment>

The expected average block time at which Subspace should be creating blocks. Since Subspace is probabilistic it is not trivial to figure out what the expected average block time should be based on the slot duration and the security parameter c (where 1 - c represents the probability of a slot being empty).

type ConfirmationDepthK: Get<Self::BlockNumber>

Depth K after which a block enters the recorded history (a global constant, as opposed to the client-dependent transaction confirmation depth k).

type ExpectedVotesPerBlock: Get<u32>

Number of votes expected per block.

This impacts solution range for votes in consensus.

type ShouldAdjustSolutionRange: Get<bool>

type GlobalRandomnessIntervalTrigger: GlobalRandomnessIntervalTrigger

Subspace requires periodic global randomness update.

type EraChangeTrigger: EraChangeTrigger

Subspace requires some logic to be triggered on every block to query for whether an era has ended and to perform the transition to the next era.

Era is normally used to update solution range used for challenges.

type HandleEquivocation: HandleEquivocation<Self>

The equivocation handling subsystem, defines methods to report an offence (after the equivocation has been validated) and for submitting a transaction to report an equivocation (from an offchain context).

NOTE: when enabling equivocation handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.